backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-rest-framework==0.1.0
django-timezone-field==5.1
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
packaging==23.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
